+++ str/str_diffn.c
@@ -5,7 +5,7 @@
  * equal to, or greater than the string b[0], b[1], ..., b[m-1]=='\0'.
  * When the strings are different, str_diff does not read bytes past the
  * first difference. */
-int str_diffn(const char* a, const char* b, unsigned int limit) {
+int str_diffn(const char* a, const char* b, unsigned long limit) {
   register const char* s=a;
   register const char* t=b;
   register const char* u=t+limit;
