+++ str/str_len.c
@@ -2,9 +2,9 @@
 
 #ifdef __dietlibc__
 #undef str_len
-unsigned int str_len(const char* in) __attribute__((alias("strlen")));
+unsigned long str_len(const char* in) __attribute__((alias("strlen")));
 #else
-unsigned int str_len(const char* in) {
+unsigned long str_len(const char* in) {
   register const char* t=in;
   for (;;) {
     if (!*t) break; ++t;
